requirements.txt
automata-lib
pandas
colormath
graphviz
jupyterlab

